requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
python-multipart
requests
beautifulsoup4
python-dotenv
psycopg2-binary
